--- .env
@@ -1,13 +1,0 @@
-DATABASE_URL=postgis://wazimap_ng:wazimap_ng@db/wazimap_ng
-DJANGO_CONFIGURATION=Local
-DJANGO_SECRET_KEY=3423424
-DJANGO_DEFAULT_FILE_STORAGE=django.core.files.storage.FileSystemStorage
-DJANGO_SERVER_INSTANCE=Development
-
-AWS_ACCESS_KEY_ID=
-AWS_SECRET_ACCESS_KEY=
-AWS_STORAGE_BUCKET_NAME=
-AWS_S3_REGION_NAME=
-
-Q_CLUSTER_WORKERS=2
-STAFF_EMAIL_ADDRESS=info@openup.org.za
