+++ requirements.txt
@@ -1,9 +1,7 @@
 # Core
 pytz==2018.9
 Django==2.2.28
-django-configurations==2.1
 gunicorn==19.9.0
-newrelic==4.12.0.113
 
 # For the persistence stores
 psycopg2-binary==2.8.5
