--- wazimap_ng/config/__init__.py
@@ -1,3 +1,0 @@
-from .local import Local  # noqa
-from .production import Production # noqa
-from .test import Test # noqa
