--- wazimap_ng/config/test.py
@@ -1,51 +1,0 @@
-import os
-import dj_database_url
-from .common import Common
-from configurations import Configuration, values
-BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-
-
-class Test(Common):
-
-    INSTALLED_APPS = Common.INSTALLED_APPS
-    SECRET_KEY = os.getenv('DJANGO_SECRET_KEY')
-
-    TEST_RUNNER = 'django_nose.NoseTestSuiteRunner'
-    NOSE_ARGS = [
-        BASE_DIR,
-        '-s',
-        # '--nologcapture',
-        # '--with-coverage',
-        # '--with-progressive',
-        '--cover-package=wazimap_ng',
-        '--cover-html'
-    ]
-
-    DATABASES = {
-        "default": dj_database_url.config(
-            default=os.getenv("DATABASE_URL", "postgis://postgres:@postgres:5432/postgres"),
-            conn_max_age=0
-        )
-    }
-
-    # Mail
-    EMAIL_HOST = 'localhost'
-    EMAIL_PORT = 1025
-    EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'
-
-    ## Honor the 'X-Forwarded-Proto' header for request.is_secure()
-    SECURE_PROXY_SSL_HEADER = ('HTTP_X_FORWARDED_PROTO', 'https')
-
-    CACHES = {
-        'default': {
-            'BACKEND': 'django.core.cache.backends.filebased.FileBasedCache',
-            'LOCATION': '/var/tmp/django_cache',
-        }
-    }
-
-    FILE_SIZE_LIMIT = 3000 * 1024 * 1024
-
-    Q_CLUSTER = {
-        "redis": os.environ.get("REDIS_URL"),
-        "sync": True
-    }
