+++ wazimap_ng/datasets/models/dataset.py
@@ -4,9 +4,7 @@ from django.contrib.postgres.fields import ArrayField
 from .geography import Geography, GeographyHierarchy, Version
 from wazimap_ng.general.models import BaseModel, SimpleHistory
 
-from wazimap_ng.config.common import (
-    PERMISSION_TYPES, DATASET_CONTENT_TYPES, QUANTITATIVE
-)
+from django.conf import settings
 
 
 class DatasetQuerySet(models.QuerySet):
@@ -22,8 +20,8 @@ class Dataset(BaseModel, SimpleHistory):
     )
     name = models.CharField(max_length=60)
     groups = ArrayField(models.CharField(max_length=200), blank=True, default=list)
-    permission_type = models.CharField(choices=PERMISSION_TYPES, max_length=32, default="private")
-    content_type = models.CharField(choices=DATASET_CONTENT_TYPES, max_length=32, default=QUANTITATIVE)
+    permission_type = models.CharField(choices=settings.PERMISSION_TYPES, max_length=32, default="private")
+    content_type = models.CharField(choices=settings.DATASET_CONTENT_TYPES, max_length=32, default=settings.QUANTITATIVE)
     version = models.ForeignKey(Version, on_delete=models.CASCADE)
 
     objects = DatasetQuerySet.as_manager()
